requirements.txt
barbar==0.2.1
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.1
cymem==2.0.3
decorator==4.4.2
en-core-web-md==2.2.5
future==0.18.2
graphviz==0.14
idna==2.9
importlib-metadata==1.6.0
joblib==0.15.1
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.5
pandas==1.0.4
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.5
pytextrank==2.0.2
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
scipy==1.4.1
six==1.15.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
torch==1.5.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
zipp==3.1.0
absl-py==0.9.0
barbar==0.2.1
blis==0.4.1
boto==2.49.0
boto3==1.14.42
botocore==1.17.42
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.20
decorator==4.4.2
docutils==0.15.2
en-core-web-md==2.2.5
future==0.18.2
gensim==3.8.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
graphviz==0.14
grpcio==1.30.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
Pillow==7.1.2
pkbar==0.4
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pyparsing==2.4.7
pytextrank==2.0.2
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scipy==1.4.1
sentencepiece==0.1.92
six==1.15.0
smart-open==2.1.0
spacy==2.2.4
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
thinc==7.4.0
torch==1.5.0
torch-two-sample==0.1
torchtext==0.6.0
torchvision==0.6.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
absl-py==0.9.0
barbar==0.2.1
blis==0.4.1
boto==2.49.0
boto3==1.14.42
botocore==1.17.42
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coverage==5.1
cycler==0.10.0
cymem==2.0.3
Cython==0.29.20
decorator==4.4.2
docutils==0.15.2
en-core-web-md==2.2.5
future==0.18.2
gensim==3.8.3
google-auth==1.19.0
google-auth-oauthlib==0.4.1
graphviz==0.14
grpcio==1.30.0
idna==2.9
importlib-metadata==1.6.0
jmespath==0.10.0
joblib==0.15.1
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.2.1
murmurhash==1.0.2
networkx==2.4
nltk==3.5
numpy==1.18.5
oauthlib==3.1.0
pandas==1.0.4
Pillow==7.1.2
pkbar==0.4
plac==1.1.3
preshed==3.0.2
protobuf==3.12.2
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pyparsing==2.4.7
pytextrank==2.0.2
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.4.1
sentencepiece==0.1.92
six==1.15.0
sklearn==0.0
smart-open==2.1.0
spacy==2.2.4
srsly==1.0.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
thinc==7.4.0
threadpoolctl==2.1.0
torch==1.5.0
torch-two-sample==0.1
torchtext==0.6.0
torchvision==0.6.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
Werkzeug==1.0.1
zipp==3.1.0
